Update minor changes as commented
